# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Internal trading platform API for Polygon market data and IBKR execution.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /algo/deployments:
    get:
      operationId: listAlgoDeployments
      tags: [automation]
      summary: List saved algo deployments
      parameters:
        - name: mode
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EnvironmentMode"
      responses:
        "200":
          description: Deployment inventory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlgoDeploymentsResponse"
    post:
      operationId: createAlgoDeployment
      tags: [automation]
      summary: Create an algo deployment linked to a promoted strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlgoDeploymentRequest"
      responses:
        "201":
          description: Created deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlgoDeployment"
  /algo/deployments/{deploymentId}/enable:
    post:
      operationId: enableAlgoDeployment
      tags: [automation]
      summary: Enable an algo deployment
      parameters:
        - name: deploymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlgoDeployment"
  /algo/deployments/{deploymentId}/pause:
    post:
      operationId: pauseAlgoDeployment
      tags: [automation]
      summary: Pause an algo deployment
      parameters:
        - name: deploymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlgoDeployment"
  /algo/events:
    get:
      operationId: listExecutionEvents
      tags: [automation]
      summary: List execution and deployment lifecycle events
      parameters:
        - name: deploymentId
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: Execution events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionEventsResponse"
  /charting/pine-scripts:
    get:
      operationId: listPineScripts
      tags: [charting]
      summary: List saved Pine scripts for chart indicators
      responses:
        "200":
          description: Pine script library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PineScriptsResponse"
    post:
      operationId: createPineScript
      tags: [charting]
      summary: Create a Pine script library entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePineScriptRequest"
      responses:
        "201":
          description: Created Pine script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PineScriptRecord"
  /charting/pine-scripts/{scriptId}:
    patch:
      operationId: updatePineScript
      tags: [charting]
      summary: Update a Pine script library entry
      parameters:
        - name: scriptId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePineScriptRequest"
      responses:
        "200":
          description: Updated Pine script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PineScriptRecord"
...
  /backtests/studies:
    get:
      operationId: listBacktestStudies
      tags: [backtests]
      summary: List saved backtest studies
      responses:
        "200":
          description: Backtest studies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestStudiesResponse"
    post:
      operationId: createBacktestStudy
      tags: [backtests]
      summary: Create a backtest study
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BacktestStudyInput"
      responses:
        "201":
          description: Created study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestStudyRecord"
  /backtests/studies/{studyId}:
    get:
      operationId: getBacktestStudy
      tags: [backtests]
      summary: Get one backtest study
      parameters:
        - name: studyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Backtest study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestStudyRecord"
  /backtests/studies/{studyId}/preview-chart:
    get:
      operationId: getBacktestStudyPreviewChart
      tags: [backtests]
      summary: Get the latest-vs-best completed study preview chart
      parameters:
        - name: studyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Study preview chart payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestStudyPreviewChart"
  /backtests/runs:
    get:
      operationId: listBacktestRuns
      tags: [backtests]
      summary: List backtest runs
      parameters:
        - name: studyId
          in: query
          required: false
          schema:
            type: string
        - name: sweepId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/BacktestJobStatus"
      responses:
        "200":
          description: Backtest runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestRunsResponse"
    post:
      operationId: createBacktestRun
      tags: [backtests]
      summary: Queue a single backtest run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBacktestRunRequest"
      responses:
        "201":
          description: Queued backtest run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestRunDetail"
  /backtests/runs/{runId}:
    get:
      operationId: getBacktestRun
      tags: [backtests]
      summary: Get one backtest run
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Backtest run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestRunDetail"
  /backtests/runs/{runId}/chart:
    get:
      operationId: getBacktestRunChart
      tags: [backtests]
      summary: Get the research-chart payload for a backtest run
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
        - name: symbol
          in: query
          required: false
          schema:
            type: string
        - name: selectedTradeId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Run chart payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestRunChart"
  /backtests/runs/{runId}/promote:
    post:
      operationId: promoteBacktestRun
      tags: [backtests]
      summary: Promote a completed run into a draft algo strategy
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoteBacktestRunRequest"
      responses:
        "201":
          description: Draft strategy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestDraftStrategy"
  /backtests/sweeps:
    post:
      operationId: createBacktestSweep
      tags: [backtests]
      summary: Queue a backtest optimizer sweep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBacktestSweepRequest"
      responses:
        "201":
          description: Queued sweep
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestSweepDetail"
  /backtests/sweeps/{sweepId}:
    get:
      operationId: getBacktestSweep
      tags: [backtests]
      summary: Get one optimizer sweep
      parameters:
        - name: sweepId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sweep detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestSweepDetail"
...
  /backtests/jobs/{jobId}/cancel:
    post:
      operationId: cancelBacktestJob
      tags: [backtests]
      summary: Request cancellation for a queued or running job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated job state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestJobSummary"
...
components:
  schemas:
...
    EnvironmentMode:
      type: string
      enum: [paper, live]
...
    BarTimeframe:
      type: string
      enum: [1s, 5s, 15s, 1m, 5m, 15m, 1h, 1d]
...
    JsonObject:
      type: object
      additionalProperties: true
...
    AlgoDeployment:
      type: object
      properties:
        id:
          type: string
        strategyId:
          type: string
        name:
          type: string
        mode:
          $ref: "#/components/schemas/EnvironmentMode"
        enabled:
          type: boolean
        providerAccountId:
          type: string
        symbolUniverse:
          type: array
          items:
            type: string
        config:
          $ref: "#/components/schemas/JsonObject"
        lastEvaluatedAt:
          type: [string, "null"]
          format: date-time
        lastSignalAt:
          type: [string, "null"]
          format: date-time
        lastError:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - strategyId
        - name
        - mode
        - enabled
        - providerAccountId
        - symbolUniverse
        - config
        - lastEvaluatedAt
        - lastSignalAt
        - lastError
        - createdAt
        - updatedAt
    CreateAlgoDeploymentRequest:
      type: object
      properties:
        strategyId:
          type: string
        name:
          type: string
        providerAccountId:
          type: string
        mode:
          $ref: "#/components/schemas/EnvironmentMode"
        symbolUniverse:
          type: array
          items:
            type: string
        config:
          $ref: "#/components/schemas/JsonObject"
      required:
        - strategyId
        - name
        - providerAccountId
        - mode
    AlgoDeploymentsResponse:
      type: object
      properties:
        deployments:
          type: array
          items:
            $ref: "#/components/schemas/AlgoDeployment"
      required:
        - deployments
...
    ExecutionEventsResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/ExecutionEvent"
      required:
        - events
...
    PineScriptStatus:
      type: string
      enum: [draft, ready, error, archived]
    PineScriptPaneType:
      type: string
      enum: [price, lower]
    PineScriptRecord:
      type: object
      properties:
        id:
          type: string
        scriptKey:
          type: string
        name:
          type: string
        description:
          type: [string, "null"]
        sourceCode:
          type: string
        status:
          $ref: "#/components/schemas/PineScriptStatus"
        defaultPaneType:
          $ref: "#/components/schemas/PineScriptPaneType"
        chartAccessEnabled:
          type: boolean
        notes:
          type: [string, "null"]
        lastError:
          type: [string, "null"]
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - scriptKey
        - name
        - description
        - sourceCode
        - status
        - defaultPaneType
        - chartAccessEnabled
        - notes
        - lastError
        - tags
        - metadata
        - createdAt
        - updatedAt
    PineScriptsResponse:
      type: object
      properties:
        scripts:
          type: array
          items:
            $ref: "#/components/schemas/PineScriptRecord"
      required:
        - scripts
    CreatePineScriptRequest:
      type: object
      properties:
        scriptKey:
          type: string
        name:
          type: string
        description:
          type: string
        sourceCode:
          type: string
        status:
          $ref: "#/components/schemas/PineScriptStatus"
        defaultPaneType:
          $ref: "#/components/schemas/PineScriptPaneType"
        chartAccessEnabled:
          type: boolean
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
      required:
        - name
        - sourceCode
    UpdatePineScriptRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        sourceCode:
          type: string
        status:
          $ref: "#/components/schemas/PineScriptStatus"
        defaultPaneType:
          $ref: "#/components/schemas/PineScriptPaneType"
        chartAccessEnabled:
          type: boolean
        notes:
          type: string
        lastError:
          type: string
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
...
    BacktestDirectionMode:
      type: string
      enum: [long_only, long_short]
    BacktestOptimizerMode:
      type: string
      enum: [grid, random, walk_forward]
    BacktestJobStatus:
      type: string
      enum:
        [
          queued,
          preparing_data,
          running,
          aggregating,
          completed,
          failed,
          cancel_requested,
          canceled,
        ]
...
    BacktestStudyInput:
      type: object
      properties:
        name:
          type: string
        strategyId:
          type: string
        strategyVersion:
          type: string
        directionMode:
          $ref: "#/components/schemas/BacktestDirectionMode"
        watchlistId:
          type: [string, "null"]
        symbols:
          type: array
          items:
            type: string
        timeframe:
          $ref: "#/components/schemas/BarTimeframe"
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        parameters:
          type: object
          additionalProperties: true
        portfolioRules:
          $ref: "#/components/schemas/BacktestPortfolioRules"
        executionProfile:
          $ref: "#/components/schemas/BacktestExecutionProfile"
        optimizerMode:
          $ref: "#/components/schemas/BacktestOptimizerMode"
        optimizerConfig:
          type: object
          additionalProperties: true
      required:
        - name
        - strategyId
        - strategyVersion
        - directionMode
        - watchlistId
        - symbols
        - timeframe
        - startsAt
        - endsAt
        - parameters
        - portfolioRules
        - executionProfile
        - optimizerMode
        - optimizerConfig
    BacktestStudyRecord:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        strategyId:
          type: string
        strategyVersion:
          type: string
        directionMode:
          $ref: "#/components/schemas/BacktestDirectionMode"
        watchlistId:
          type: [string, "null"]
        symbols:
          type: array
          items:
            type: string
        timeframe:
          $ref: "#/components/schemas/BarTimeframe"
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        parameters:
          type: object
          additionalProperties: true
        portfolioRules:
          $ref: "#/components/schemas/BacktestPortfolioRules"
        executionProfile:
          $ref: "#/components/schemas/BacktestExecutionProfile"
        optimizerMode:
          $ref: "#/components/schemas/BacktestOptimizerMode"
        optimizerConfig:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - strategyId
        - strategyVersion
        - directionMode
        - watchlistId
        - symbols
        - timeframe
        - startsAt
        - endsAt
        - parameters
        - portfolioRules
        - executionProfile
        - optimizerMode
        - optimizerConfig
        - createdAt
        - updatedAt
    CreateBacktestRunRequest:
      type: object
      properties:
        studyId:
          type: string
        name:
          type: [string, "null"]
        parameters:
          type: [object, "null"]
          additionalProperties: true
      required:
        - studyId
        - name
        - parameters
...
    CreateBacktestSweepRequest:
      type: object
      properties:
        studyId:
          type: string
        mode:
          $ref: "#/components/schemas/BacktestOptimizerMode"
        baseParameters:
          type: object
          additionalProperties: true
        dimensions:
          type: array
          items:
            $ref: "#/components/schemas/BacktestSweepDimension"
        randomCandidateBudget:
          type: [integer, "null"]
        walkForwardTrainingMonths:
          type: [integer, "null"]
        walkForwardTestMonths:
          type: [integer, "null"]
        walkForwardStepMonths:
          type: [integer, "null"]
      required:
        - studyId
        - mode
        - baseParameters
        - dimensions
        - randomCandidateBudget
        - walkForwardTrainingMonths
        - walkForwardTestMonths
        - walkForwardStepMonths
...
    BacktestRunChart:
      type: object
      properties:
        runId:
          type: string
        studyId:
          type: string
        timeframe:
          $ref: "#/components/schemas/BarTimeframe"
        chartPriceContext:
          type: string
          enum: ["spot", "option"]
        availableSymbols:
          type: array
          items:
            type: string
        selectedSymbol:
          type: string
        defaultTradeSelectionId:
          type: [string, "null"]
        activeTradeSelectionId:
          type: [string, "null"]
        chartBars:
          type: array
          items:
            $ref: "#/components/schemas/BacktestChartBar"
        chartBarRanges:
          type: array
          items:
            $ref: "#/components/schemas/BacktestChartBarRange"
        tradeOverlays:
          type: array
          items:
            $ref: "#/components/schemas/BacktestTradeOverlay"
        tradeMarkerGroups:
          $ref: "#/components/schemas/BacktestTradeMarkerGroups"
        indicatorEvents:
          type: array
          items:
            $ref: "#/components/schemas/BacktestIndicatorEvent"
        indicatorZones:
          type: array
          items:
            $ref: "#/components/schemas/BacktestIndicatorZone"
        indicatorWindows:
          type: array
          items:
            $ref: "#/components/schemas/BacktestIndicatorWindow"
        indicatorMarkerPayload:
          $ref: "#/components/schemas/BacktestIndicatorMarkerPayload"
        selectionFocus:
          oneOf:
            - $ref: "#/components/schemas/BacktestTradeSelectionFocus"
            - type: "null"
        defaultVisibleLogicalRange:
          oneOf:
            - type: object
              properties:
                from:
                  type: number
                to:
                  type: number
              required:
                - from
                - to
            - type: "null"
      required:
        - runId
        - studyId
        - timeframe
        - chartPriceContext
        - availableSymbols
        - selectedSymbol
        - defaultTradeSelectionId
        - activeTradeSelectionId
        - chartBars
        - chartBarRanges
        - tradeOverlays
        - tradeMarkerGroups
        - indicatorEvents
        - indicatorZones
        - indicatorWindows
        - indicatorMarkerPayload
        - selectionFocus
        - defaultVisibleLogicalRange
...
    BacktestStudyPreviewChart:
      type: object
      properties:
        studyId:
          type: string
        latestCompletedRun:
          oneOf:
            - $ref: "#/components/schemas/BacktestRunSummary"
            - type: "null"
        bestCompletedRun:
          oneOf:
            - $ref: "#/components/schemas/BacktestRunSummary"
            - type: "null"
        comparisonBadges:
          type: array
          items:
            $ref: "#/components/schemas/BacktestComparisonBadge"
        latestSeries:
          type: array
          items:
            $ref: "#/components/schemas/BacktestEquitySeriesPoint"
        bestSeries:
          type: array
          items:
            $ref: "#/components/schemas/BacktestEquitySeriesPoint"
      required:
        - studyId
        - latestCompletedRun
        - bestCompletedRun
        - comparisonBadges
        - latestSeries
        - bestSeries
    BacktestRunDetail:
      type: object
      properties:
        run:
          $ref: "#/components/schemas/BacktestRunSummary"
        study:
          $ref: "#/components/schemas/BacktestStudyRecord"
        trades:
          type: array
          items:
            $ref: "#/components/schemas/BacktestTrade"
        points:
          type: array
          items:
            $ref: "#/components/schemas/BacktestPoint"
        datasets:
          type: array
          items:
            $ref: "#/components/schemas/BacktestDatasetRef"
      required:
        - run
        - study
        - trades
        - points
        - datasets
    BacktestSweepDetail:
      type: object
      properties:
        id:
          type: string
        studyId:
          type: string
        mode:
          $ref: "#/components/schemas/BacktestOptimizerMode"
        status:
          $ref: "#/components/schemas/BacktestJobStatus"
        candidateTargetCount:
          type: integer
        candidateCompletedCount:
          type: integer
        bestRunId:
          type: [string, "null"]
        startedAt:
          type: [string, "null"]
          format: date-time
        finishedAt:
          type: [string, "null"]
          format: date-time
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/BacktestRunSummary"
      required:
        - id
        - studyId
        - mode
        - status
        - candidateTargetCount
        - candidateCompletedCount
        - bestRunId
        - startedAt
        - finishedAt
        - candidates
...
    PromoteBacktestRunRequest:
      type: object
      properties:
        name:
          type: string
        notes:
          type: [string, "null"]
      required:
        - name
        - notes
    BacktestDraftStrategy:
      type: object
      properties:
        id:
          type: string
        runId:
          type: string
        studyId:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        mode:
          $ref: "#/components/schemas/EnvironmentMode"
        symbolUniverse:
          type: array
          items:
            type: string
        config:
          type: object
          additionalProperties: true
        promotedAt:
          type: string
          format: date-time
      required:
        - id
        - runId
        - studyId
        - name
        - enabled
        - mode
        - symbolUniverse
        - config
        - promotedAt
...
    BacktestStudiesResponse:
      type: object
      properties:
        studies:
          type: array
          items:
            $ref: "#/components/schemas/BacktestStudyRecord"
      required:
        - studies
    BacktestRunsResponse:
      type: object
      properties:
        runs:
          type: array
          items:
            $ref: "#/components/schemas/BacktestRunSummary"
      required:
        - runs
...